requirements.txt
scikit-learn
pandas
apache-airflow